backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
